Criteria Helper Functions for Custom Search

Criteria helper functions to be combined inside a operator helper function as a custom request in customSearch.

String(section_or_field, string, negate = FALSE)

Arguments

section_or_field
A mandatory character string specifying in which messages's Section or Header Field to search for the provided string. For some available options, see section_or_field_options.

string
A character string specifying the word or expression to search for in messages.

negate
If TRUE, negates the search and seeks for "NOT search_criterion". Default is FALSE.

Value

A search string to be used as a custom_request parameter in customSearch function.

See also

Examples

# NOT RUN {
# configure IMAP
library(mRpostman)
imapconf <- configureIMAP(url="imaps://imap.gmail.com",
                          username="your_gmail_user",
                          password=rstudioapi::askForPassword()
                          )

# search
result <- imapconf %>%
    selectMailbox(mbox = "INBOX") %>%
    customSearch(custom_request =
                 OR(String(
                       section_or_field = "from", string = "allanvcq@gmail.com"),
                    String(
                       section_or_field = "from", string = "allanvcq@yahoo.com")
                   )
                 )
# searches for messages containing the string "allanvcq@gmail.com" OR
    "allanvcq@yahoo.com" in the "FROM" field.

# }