Metadata Fetch Arguments Check

Internal helper function for checking arguments used in fetch functions.

check_args_fetch_msg_metadata(imapconf, msg_id, by, metadata,
  write_to_file, keep_in_mem, retries)

Arguments

imapconf
An object of class imapconf generated with configure_imap and with a mbox item added after a mailbox selection with select_mailbox.

msg_id
A numeric vetor containing one or more messages ids.

by
Which id (MSN or UID) to use when searching for messages. "MSN" or message squence number is a message's relative position to the older message in the mailbox. It may change after deleting or moving messages. For instance, if a message is deleted, message sequence numbers are reordered to fill the gap. "UID" or unique identifier is always the same during the life cycle of a message. Default is "MSN".

metadata
A character vector or string specifying one or more items of the metadata of a message to fetch. See metadata_options.

write_to_file
If TRUE writes fetch content of each message to the disk as a text file in the working directory. Default is FALSE.

keep_in_mem
If TRUE keeps a copy of fetch results as an list in the R session when write_to_file = TRUE. Default is FALSE. It can only be set TRUE when write_to_file = TRUE.

retries
Number of attempts to connect and execute the command. Default is 2.

Value

NULL if arguments are correct.

See also