Size Search Request

Internal helper function for defining curl custom request to be used with each search function

define_searchrequest_size(operation, size, negate, by, flag, esearch,
  handle)

Arguments

operation
A character indicating which custom operation to execute.

size
An integer specifying the size (in number of characters) of message to be used as search criterion.

negate
If TRUE, negates the search and seeks for "NOT search_criterion". Default is FALSE.

by
Which id (MSN or UID) to use when searching for messages. "MSN" or message squence number is a message's relative position to the older message in the mailbox. It may change after deleting or moving messages. For instance, if a message is deleted, message sequence numbers are reordered to fill the gap. "UID" or unique identifier is always the same during the life cycle of a message. Default is "MSN".

flag
Optional parameter that adds a flag filter to the search. Use flag_options to list the common flags used by IMAP servers. Default is NULL.

esearch
A logical. Default is FALSE. If the IMAP server has ESEARCH capability, it can be used to optimize search results. It allows to condense results to message sets to cut down on transmission costs, e.g. 1:5 instead of writing ids individually such as code1,2,3,4,5. It can be used along with buffersize to avoid results stripping. Check if your IMAP server supports SEARCH with listServerCapabilities.

handle
A curl handle object.

Value

A curl handle object containing the custom request.

See also