Execution engine loop for all the fetch commands

Execution engine loop for all the fetch commands

execute_fetch_loop(
  self,
  msg_id,
  fetch_request,
  use_uid,
  write_to_disk,
  keep_in_mem,
  retries,
  fetch_type,
  base64_decode = FALSE,
  metadata_attribute = NULL
)

Arguments

self
The R6 connection object.

msg_id
A numeric vector containing one or more message ids.

fetch_request
A string containing the fetch request to the server that will be added to the curl handle.

use_uid
Default is FALSE. In this case, results will be presented as message's sequence numbers. A message sequence number is a message's relative position to the oldest message in the mailbox. It may change after deleting or moving messages. If a message is deleted, sequence numbers are reordered to fill the gap. If TRUE, the command will be performed using the "UID" or unique identifier, and results are presented as such. UIDs are always the same during the life cycle of a message.

write_to_disk
If TRUE, writes the fetch content of each message to a text file in a local folder inside the working directory, also returning the results with invisible(). Default is FALSE.

keep_in_mem
If TRUE, keeps a copy of each fetch result while the operation is being performed with write_to_disk = TRUE. Default is FALSE, and it can only be set TRUE when write_to_disk = TRUE.

retries
Number of attempts to connect and execute the command. Default is 1.

fetch_type
A string indicating if it will be executed a body, header, text, or metadata fetch.

base64_decode
If TRUE, tries to guess and decode the fetched text from base64 format to character. Default is FALSE. Only used in the fetch_text() case.

metadata_attribute
String containing the meta attributes if applicable. Default is NULL.

@noRd