Search By Internal Date

Functions that allows searching for messages using the (internal) date criteria, such as before, since, on, and period.

search_before(imapconf, date_char, negate = FALSE, by = "MSN",
  flag = NULL, esearch = FALSE, return_imapconf = TRUE,
  retries = 2)

Arguments

imapconf
An object of class imapconf generated with configure_imap and with a mbox item added after a mailbox selection with select_mailbox.

date_char
A character vector with format "DD-Mon-YYYY", e.g. "01-Apr-2019". We opted not to use objects of type "date", since IMAP servers like this not so common date format.

negate
If TRUE, negates the search and seeks for "NOT search_criterion". Default is FALSE.

by
Which id (MSN or UID) to use when searching for messages. "MSN" or message squence number is a message's relative position to the older message in the mailbox. It may change after deleting or moving messages. For instance, if a message is deleted, message sequence numbers are reordered to fill the gap. "UID" or unique identifier is always the same during the life cycle of a message. Default is "MSN".

flag
Optional parameter that adds a flag filter to the search. Use flag_options to list the common flags used by IMAP servers. Default is NULL.

esearch
A logical. Default is FALSE. If the IMAP server has ESEARCH capability, it can be used to optimize search results. It allows to condense results to message sets to cut down on transmission costs, e.g. 1:5 instead of writing ids individually such as code1,2,3,4,5. It can be used along with buffersize to avoid results stripping. Check if your IMAP server supports SEARCH with list_server_capabilities.

return_imapconf
A logical. If TRUE, the function returns a list of length 2, containing the imapconf object (IMAP settings) and the search results as message ids ("MSN" or "UID"). If FALSE, returns only the message ids as a numeric vector. Default is TRUE.

retries
Number of attempts to connect and execute the command. Default is 2.

Value

Depending on the return_imapconf parameter, it can be a list of length 2 containing the imapconf object and message ids resulted from search, or a numeric vector containing only message ids.

Note

search_before, search_since, search_on, and search_period use internal date, which reflects the moment when the message was received. search_sent_before, search_sent_since, search_sent_on, and search_sent_period use RFC-2822 date header (origination date), which "specifies the date and time at which the creator of the message indicated that the message was complete and ready to enter the mail delivery system" (Resnick, 2008). Dates in both methods must be the same most of time. Nonetheless, using internal date for search is faster (Babcock, 2016).

References

Resnick, P., Ed., "Internet Message Format", RFC 5322, doi: 10.17487/RFC5322 , October 2008, https://www.rfc-editor.org/info/rfc5322.

Babcock, N., "Introduction to IMAP", Blog, May 2016, https://nbsoftsolutions.com/blog/introduction-to-imap.

See also

Examples

# NOT RUN {
# configure IMAP
library(mRpostman)
imapconf <- configure_imap(url="imaps://imap.gmail.com",
                          username="your_gmail_user",
                          password=rstudioapi::askForPassword()
                          )

# search 1
result1 <- imapconf %>%
    select_mailbox(mbox = "[Gmail]/Sent") %>%
    search_before(date_char = "17-Apr-2012")

# search 2 (brazilian portuguese example -- mailboxes with compound names)
result2 <- imapconf %>%
    select_mailbox(mbox = "[Gmail]/Com estrela") %>%
    search_before(date_char = "17-Dec-2018", flag = "ANSWERED")
# depending on results, it might be necessary increase buffersize in configureIMAP()

# }