Criteria Helper Functions for Custom Search

Criteria helper functions to be combined inside a operator helper function as a custom request in customSearch.

sentOn(date_char, negate = FALSE)

Arguments

date_char
A character vector with format "DD-Mon-YYYY", e.g. "01-Apr-2019". We opted not to use objects of type "date", since IMAP servers like this not so common date format.

negate
If TRUE, negates the search and seeks for "NOT search_criterion". Default is FALSE.

Value

A search string to be used as a custom_request parameter in customSearch function.

See also

Examples

# NOT RUN {
# configure IMAP
library(mRpostman)
imapconf <- configureIMAP(url="imaps://imap.gmail.com",
                          username="your_gmail_user",
                          password=rstudioapi::askForPassword()
                          )

# search
result <- imapconf %>%
    selectMailbox(mbox = "INBOX") %>%
    customSearch(custom_request = OR(sentOn(date_char = "17-Apr-2019"),
                                     sentOn(date_char = "17-Jul-2018")))
# searches for messages sentOn "17-Apr-2019" OR sentOn "17-Jul-2018".

# }